[config] Fix handling of hidden modes setting
The make_hidden_modes_string() function claims to return const string,
but that is true only as an exception to the rule. Since the callers of
the function do not release dynamic memory, this leads to leakage. Also
if no mode is hidden, request to unhide something least to that mode
getting hidden and hiding something already hidden returns NULL which
is rather unintuitive.

Fix make_hidden_modes_string() and make all callers of it to release
dynamically reserved memory / deal with possible null return value.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>