[trigger] Fix udev iowatch issues
Various pointers are uninitialized / left to hold stale values after
releasing. Which could lead to hard to debug problems if trigger_stop() is
called multiple times / after unsuccessful trigger_init().

Zero initialize all pointers dealing with dynamic resources and clear them
when resources are released.

Also clear iowatch id when it is going to be implicitly removed due to the
return value from the callback so that removal is not attempted anymore
when/if trigger_stop() is called later on.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>