[modesetting] On debug verbosity log all sysfs writes made
When debugging mode setting issues, it is hard to tell when and what gets
written to android usb sysfs files.

When running usb-moded in debug verbosity, log sysfs paths that get written
to and the data that is written - and if possible also the value held by
the sysfs file before the write.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>